circleseq/requirements.txt
HTSeq==0.6.1p1
PyYAML==3.11
swalign==0.3.1
pyfaidx==0.2.7
svgwrite==1.1.6
biopython==1.68
pandas==